closes #20 memory leak issue
Resolved:
```
sh build_all.sh && valgrind ./bin/time_tracker.cpp.bin
==19879== Memcheck, a memory error detector
==19879== Copyright (C) 2002-2011, and GNU GPL'd, by Julian Seward et al.
==19879== Using Valgrind-3.7.0 and LibVEX; rerun with -h for copyright info
==19879== Command: ./bin/time_tracker.cpp.bin
==19879==
[INFO] [26/01/2013 11:40:47.187] Result is 3
==19879==
==19879== HEAP SUMMARY:
==19879==     in use at exit: 0 bytes in 0 blocks
==19879==   total heap usage: 109 allocs, 109 frees, 44,853 bytes allocated
==19879==
==19879== All heap blocks were freed -- no leaks are possible
==19879==
==19879== For counts of detected and suppressed errors, rerun with: -v
==19879== ERROR SUMMARY: 0 errors from 0 contexts (suppressed: 2 from 2)
```

+++ update.sh
@@ -36,6 +36,14 @@ fi
 echo "Current File : $currFile"
 echo "targetFile : $targetFile"
 echo "newVersionFile : $newVersionFile"
+echo
+echo "** NOTE: THIS IS UNSTABLE VERSION OF update.sh, PLEASE REFER TO ISSUE #12 ON GITHUB AT github.com/mkhan3189/EasyLoggingPP/issues/12 FOR DETAILS **"
+echo
+echo "Continue? (y/n)"
+read cont
+if [ "$cont" = "n" ];then
+  exit
+fi
 
 [ -f "$targetFile" ] || rm $targetFile
